+++ include/utils.hpp
@@ -7,19 +7,19 @@
 
 namespace Utils {
     // Get module name which includes given address.
-    std::string GetModuleNameByAddr(DWORD address);
+    std::string GetModuleNameByAddr(size_t address);
     // Get symbol name of given address.
-    std::string GetSymbolName(DWORD called);
+    std::string GetSymbolName(size_t called);
 
     // Set trap flag.
     void SetSingleStep(PCONTEXT context);
     // Set software breakpoint.
-    bool SetBreakPoint(DWORD address);
+    bool SetBreakPoint(size_t address);
     // Release breakpoint.
-    bool RecoverBreakpoint(DWORD address);
+    bool RecoverBreakpoint(size_t address);
     
     // Get text section address.
-    std::tuple<DWORD, DWORD> GetTextSectionAddress();
+    std::tuple<size_t, size_t> GetTextSectionAddress();
 }
 
 #endif
